+++ bin/happy_hour_server.dart
@@ -4,6 +4,7 @@ import 'package:happy_hour_server/server/runner.dart';
 import 'package:happy_hour_server/server/server.dart';
 
 import 'api/register.dart';
+import 'api/v1.dart';
 import 'api/v1/auction.dart';
 import 'api/v1/auctions.dart';
 import 'api/v1/auth.dart';
@@ -20,6 +21,7 @@ class MyServer extends Server {
 		router.addRoute('/api/v1/user', filters: [FilterLogged()], dispatcher: ApiV1User());
 		router.addRoute('/api/v1/auth', filters: [FilterLogged()], dispatcher: ApiV1Auth());
 		router.addRoute('/api/register', dispatcher: ApiRegister());
+		router.addRoute('/api/v1', filters: [FilterLogged()], dispatcher: ApiV1());
 	}
 
 }